test(quick-n70): split hypermutated tests by source module
Split test_pcgr_hypermutated.py into test_pcgr.py and
test_smlv_somatic_report.py, matching bolt.common.pcgr and
bolt.workflows.smlv_somatic.report; shared VCF fixtures moved to
tests/helpers.py.
